+++ blueprint/interfaces/billing/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @forklaunch/interfaces-billing
 
+## 0.8.19
+
+### Patch Changes
+
+- Package bumps
+
 ## 0.8.18
 
 ### Patch Changes