+++ framework/express/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @forklaunch/express
 
+## 0.10.12
+
+### Patch Changes
+
+- Downgrade mikro-orm back to normal
+- Updated dependencies
+  - @forklaunch/validator@0.10.35
+  - @forklaunch/common@0.6.35
+  - @forklaunch/core@0.18.8
+
+## 0.10.11
+
+### Patch Changes
+
+- bump packages and internal proxy await resilience
+- Updated dependencies
+  - @forklaunch/validator@0.10.34
+  - @forklaunch/common@0.6.34
+  - @forklaunch/core@0.18.7
+
 ## 0.10.10
 
 ### Patch Changes